% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{cde_and_mass_lei_wasserman_discrete}
\alias{cde_and_mass_lei_wasserman_discrete}
\title{Caculate CDE Y|X=x from data example from in Lei and Wasseerman (2014) paper
on a grid of y values}
\usage{
cde_and_mass_lei_wasserman_discrete(
  x,
  grid = seq(-8, 8, length.out = 100),
  verbose = T
)
}
\arguments{
\item{x}{x value (can be a vector)}

\item{grid}{vector of y values to be evaluated (must be equally spaced)}

\item{verbose}{boolean, if we should enumerate the how many \code{x} values
it has examined values for.}
}
\value{
data frame with columns
\itemize{
\item{\code{x}: }{the associated x values}
\item{\code{y}: }{the associated y values}
\item{\code{cde}: }{CDE for Y=\code{y}|X=\code{x} as described above.}
\item{\code{mass}: }{Approximation of P(z|x and f(z|x) >= f(y|x))}
}
}
\description{
The generative model is defined as
\deqn{(Y|X=x) ~ .5 N(f(x)-g(x), sigma^2(x)) + .5 N(f(x)+g(x),
     sigma^2(x))}{(Y|X=x) \sim .5 N(f(x)-g(x),
     \sigma^2(x)) + .5 N(f(x)+g(x),\sigma^2(x))}
where
\enumerate{
\item \eqn{f(x) = (x-1)^2(x+1)},
\item \eqn{g(x) = 2 sqrt(x+.5) II(x >= -.5)}{
    g(x) = 2\sqrt(x+.5)\mathbb{I}(x\geq -.5)}, and
\item \eqn{sigma(x) = 1/4 + |x|.}{\sigma(x) = 1/4 + |x|.}
}
}
