% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{cde_lei_wassserman}
\alias{cde_lei_wassserman}
\title{Create the CDE Y|X=x for data example from in Lei and Wasseerman (2014) paper}
\usage{
cde_lei_wassserman(x)
}
\arguments{
\item{x}{x value (a scalar)}
}
\value{
function to calculate the CDE for any y value
}
\description{
The generative model is defined as
\deqn{(Y|X=x) ~ .5 N(f(x)-g(x), sigma^2(x)) + .5 N(f(x)+g(x),
     sigma^2(x))}{(Y|X=x) \sim .5 N(f(x)-g(x),
     \sigma^2(x)) + .5 N(f(x)+g(x),\sigma^2(x))}
where
\enumerate{
\item \eqn{f(x) = (x-1)^2(x+1)},
\item \eqn{g(x) = 2 sqrt(x+.5) II(x >= -.5)}{
    g(x) = 2\sqrt(x+.5)\mathbb{I}(x\geq -.5)}, and
\item \eqn{sigma(x) = 1/4 + |x|.}{\sigma(x) = 1/4 + |x|.}
}
}
