% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_exists}
\alias{check_file_exists}
\title{check if a file exists with a certain name in a certain directory}
\usage{
check_file_exists(file_name, dir = ".", .logic = T)
}
\arguments{
\item{file_name}{file name or part of string name (if not unique will just
check that at least a single file matches the provided name)}

\item{dir}{file directory (default is current directory)}

\item{.logic}{boolean - if should return a boolean (\code{TRUE}) or should
return a vector a files that match \code{file_name} string (\code{FALSE}).}
}
\value{
see \code{.logic} parameter.
}
\description{
Uses \code{stringr::str_detect}, if you want exact, use \code{^} and \code{$}
before and after your string
}
