% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_pdf_approach.R
\name{conformal_pdf_breaks}
\alias{conformal_pdf_breaks}
\title{calculate conformal scores and split into level groupings}
\usage{
conformal_pdf_breaks(
  individual_df,
  split_conformal_info_vector,
  conformal_score = c("cde", "mass")
)
}
\arguments{
\item{individual_df}{data.frame that contains information from a single (x)
with the following columns:
\itemize{
\item{\code{yy}:}{ range of y value that conformal inference will explore}
\item{\code{fit_density}:}{ fit density estimates for \code{yy}}
}}

\item{split_conformal_info_vector}{a vector of conformal / non-conformal
scores from a calibration set}

\item{conformal_score}{string to indicate if we will be doing use the CDE
conformal score (\code{"cde"}) or mass non-conformal score (\code{"mass"})}
}
\value{
\code{individual_df} with additional columns, specifically:
\itemize{
\item{\code{g_id}:}{ string defining which conformal "cut" of the conformal
score values the fit density lives in}
\item{\code{g_id2}:}{ integer factor representation of \code{g_id}}
\item{\code{g_id3}:}{ \code{g_id2}/\eqn{max}(\code{g_id2})}
\item{\code{cut_off_lower}:}{ lower threshold from \code{g_id} as scalar}
\item{\code{cut_off_upper}:}{ upper threshold from \code{g_id} as scalar}
}
}
\description{
calculate conformal scores and split into level groupings
}
\details{
For \code{split_conformal_info_vector} need to feed in acceptable conformal
or non-conformal scores depending upon \code{conformal_score} string value.

In the output, the "cuts" are assoicated with
\eqn{y:\hat{f}(y|x) > \lambda_i}
}
