% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{conformal_scores_lei_wasserman}
\alias{conformal_scores_lei_wasserman}
\title{calculate true CDE conformal score and approximate mass non-conformal score
under Lei & Wasserman 2014 perfect data fit.}
\usage{
conformal_scores_lei_wasserman(x, y, grid = seq(-8, 8, length.out = 100))
}
\arguments{
\item{x}{x value}

\item{y}{observed y value}

\item{grid}{good range of potential y values that take up most of mass of
CDE, should also be equally spaced.}
}
\value{
data frame with
\itemize{
\item{\code{x}: }
\item{\code{y}: }
\item{\code{cde}: }{CDE of true y value}
\item{\code{mass}: }{approximation of P(z|x and f(z|x) >= f(y|x))}
}
}
\description{
calculate true CDE conformal score and approximate mass non-conformal score
under Lei & Wasserman 2014 perfect data fit.
}
