% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{coverage_down_1d_mult}
\alias{coverage_down_1d_mult}
\title{Calculates the minimum coverage radius for euclidean objects as we increase
the number of observations relative to mode groupings.}
\usage{
coverage_down_1d_mult(sim_df, e_cols, ordering_list, verbose = FALSE)
}
\arguments{
\item{sim_df}{data frame of points as rows}

\item{e_cols}{character vector of column names that define
locational information of points}

\item{ordering_list}{list of indices, each vector is related to a single mode
cluster and the ordering defines how we build up the region.}

\item{verbose}{boolean, if this progression should be verbose}
}
\value{
list per mode grouping in \code{ordering_list} with inner lists of
min_cover_vec (minimum at that step) and dist_mat (cumulative
max per path at time step t (column))
}
\description{
Calculates the minimum coverage radius for euclidean objects as we increase
the number of observations relative to mode groupings.
}
