% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{coverage_down_1d_single}
\alias{coverage_down_1d_single}
\title{Calculates the minimum coverage radius for euclidean objects as we increase
the number of observations.}
\usage{
coverage_down_1d_single(ordered_sim_df, e_cols, verbose = FALSE)
}
\arguments{
\item{ordered_sim_df}{data frame of points as rows, ordered by when they
should be added in the sequence}

\item{e_cols}{character vector of column names that define
locational information of points}

\item{verbose}{boolean, if this progression should be verbose}
}
\value{
list of min_cover_vec (minimum at that step) and dist_mat (cumulative
max per path at time step t (column))
}
\description{
Calculates the minimum coverage radius for euclidean objects as we increase
the number of observations.
}
