% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_pdf_approach.R
\name{create_S_b_matrix}
\alias{create_S_b_matrix}
\title{Create the "S" matrix and "b" vector}
\usage{
create_S_b_matrix(n, prob_cde, prime = FALSE)
}
\arguments{
\item{n}{number of observations in the calibration set}

\item{prob_cde}{\eqn{P(y: f(y|x) \in [lambda_{i}, lambda_{i+1})} vector. See
details for more information.}

\item{prime}{boolean if we should return S' and b'}
}
\value{
list of S matrix and b vector.
}
\description{
An internal function that create the "S" matrix and "b" vector for the
"conformal pdf".
}
\details{
The S matrix is a the upper triangle, where the non-zero values have column
information that is \eqn{P(y: f(y|x) \in [lambda_{i}, lambda_{i+1})} where
\eqn{\lambda_{i}} are the level set ups associated with conformal scores
values.

The b vector is just (i-1)/n
}
