% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crossdist}
\alias{crossdist}
\title{calculates distances between points of 2 matrices}
\usage{
crossdist(m1, m2)
}
\arguments{
\item{m1}{first matrix (n x p)}

\item{m2}{second matrix (m x p)}
}
\description{
\code{C++} function taken from
https://www.r-bloggers.com/2019/09/matrix-cross-distances-using-rcpp/
}
