% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{crossdist_df}
\alias{crossdist_df}
\title{euclidean distance between 2 sets}
\usage{
crossdist_df(df1, df2)
}
\arguments{
\item{df1}{data.frame (n x p)}

\item{df2}{data.frame (m x p)}
}
\value{
distance between x points (rows) and y points (columns) - (n x m)
}
\description{
euclidean distance between 2 sets
}
