% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_pdf_approach.R
\name{cumulative_comparisons}
\alias{cumulative_comparisons}
\title{calculate cumulative comparisons}
\usage{
cumulative_comparisons(
  individual_df_rich,
  group_columns = c("g_id", "g_id2", "g_id3", "cut_off_lower", "cut_off_upper"),
  delta_y = NULL
)
}
\arguments{
\item{individual_df_rich}{data.frame that contains information from a single (x)
with the following columns:
\itemize{
\item{\code{yy}:}{ range of y value that conformal inference will explore}
\item{\code{fit_density}:}{ fit density estimates for \code{yy}}
}
and also contains conformal grouping structure (identified in
\code{group_columns}) and that relative to conformal level groupings}

\item{group_columns}{tidyverse style column groupings (all columns that
identify information about the conformal level groupings)}

\item{delta_y}{difference betwen \code{yy} values (assumed constant), the
default is to calculate them internally.}
}
\value{
\code{cumulative_df} contains cumulative mass information for the
fit density and true density (currently assumed known...)
}
\description{
calculate cumulative comparisons
}
