% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cut_to_numeric}
\alias{cut_to_numeric}
\title{converts interval values from the \code{cut} function to scalars. Works
inside \code{tidyverse}'s \code{mutate} function.}
\usage{
cut_to_numeric(id, .lower = TRUE)
}
\arguments{
\item{id}{value or set of values}

\item{.lower}{boolean to isolate the lower bound or upper bound. Default is
\code{TRUE} - aka .lower bound.}
}
\value{
vector of scalar values for the bound
}
\description{
converts interval values from the \code{cut} function to scalars. Works
inside \code{tidyverse}'s \code{mutate} function.
}
