% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{distdex}
\alias{distdex}
\title{Identify the index in a raveled \code{dist} (upper triangle of distance)
relative to row / column ordering}
\usage{
distdex(i, j, n, .check_order = T, .check_possible_ij = T)
}
\arguments{
\item{i}{row index (either a single value or vector). \code{i} < \code{j}.}

\item{j}{column index (either a single value or vector)}

\item{n}{number of objects that created the \code{dist} object}

\item{.check_order}{boolean (if we should check that \code{i} < \code{j})
and remove all values that don't follow that rule from returning.}

\item{.check_possible_ij}{boolean (if we should check that \code{i} and
\code{j} are not greater than \code{n}), and remove all values that don't
follow that rule from returning.}
}
\value{
requested indicies.
}
\description{
A extended idea as proposed on
\href{stack overflow}{https://stackoverflow.com/a/12643509} by Christian A.
}
\details{
Note that i and j can be vectors, but if they're both vectors we
don't return all combinations of each.
}
