% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{distEuclidean}
\alias{distEuclidean}
\title{euclidean distance between 2 sets}
\usage{
distEuclidean(x, y)
}
\arguments{
\item{x}{data.frame (n x p)}

\item{y}{data.frame (m x p)}
}
\value{
distance between x points (rows) and y points (columns) - (n x m)
}
\description{
euclidean distance between 2 sets
}
