% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulationBands-package.R
\docType{package}
\name{EpiCompare}
\alias{EpiCompare}
\title{simulationBands: package overview}
\description{
The goal of simulationBands is to provide a set of tools to implement
simulation-based conformal inference for different applications.
}
