% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Y.R
\name{gen_mix_reg}
\alias{gen_mix_reg}
\title{generate 1d "v" distribution from x values}
\usage{
gen_mix_reg(
  x,
  beta = c(-1, 1),
  intercept = c(0, 0),
  sd = c(1, 1),
  proportions = c(0.5, 0.5)
)
}
\arguments{
\item{x}{x values}

\item{beta}{vector length 2, see above}

\item{intercept}{vector length 2, see above}

\item{sd}{vector length 2, see above}

\item{proportions}{vector length 2, see above}
}
\value{
data frame (x,y) sampled from the above distribution
}
\description{
Data is generated from y ~ p1 M1(x) + p2 M2(x)
where (M1,M2)(x) ~ right_intercept + N(right_beta*x, right_sd) and
(p1,p2) = right_proportions
}
