% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{get_delta}
\alias{get_delta}
\alias{get_delta_simple}
\title{calculate maxmin distance between points}
\usage{
get_delta(dist_mat)

get_delta_simple(dist_mat)
}
\arguments{
\item{dist_mat}{a symmetric matrix with distance values, captures the
distance col observation is from row observation (digonals are expected to be
0).}
}
\value{
minimum radius for all points to be covered
}
\description{
Can be though of as a LOOCV containment of all points (what is the minimum
radius to do so).
}
