% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{get_delta_dist}
\alias{get_delta_dist}
\title{calculate maxmin distance between points (\code{dist} based)}
\usage{
get_delta_dist(x, verbose = F)
}
\arguments{
\item{x}{\code{dist} based object}

\item{verbose}{boolean, if should present a progress bar}
}
\value{
minimum radius for all points to be covered
}
\description{
Can be though of as a LOOCV containment of all points (what is the minimum
radius to do so).
}
