% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{get_delta_flex}
\alias{get_delta_flex}
\title{calculate maxmin distance between points}
\usage{
get_delta_flex(df, full_breaks = 10, verbose = F)
}
\arguments{
\item{df}{data.frame (with only columns that are needed)}

\item{full_breaks}{number of breaks relative to \code{get_delta_large},
otherwise it's not used.}

\item{verbose}{boolean, a progressbar is provided  to track the progress of
\code{get_delta_dist} or \code{get_delta_large}, if they are used.}
}
\value{
minimum radius for all points to be covered
}
\description{
A wrapper of \code{get_delta_simple}, \code{get_delta_dist}, and
\code{get_delta_large} relative to the number of observations to be used.
Attempts to balance the number of observations vs the fastest approach (
attempting to balance time costs from memory storage and processing speed).
}
