% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{get_delta_large}
\alias{get_delta_large}
\title{calculate maxmin distance between points (breaks into parts)}
\usage{
get_delta_large(df, full_breaks = 10, verbose = F)
}
\arguments{
\item{df}{data.frame}

\item{full_breaks}{number of breaks (1 more if not perfectly split)}

\item{verbose}{boolean, if should present a progress bar}
}
\value{
minimum radius for all points to be covered
}
\description{
Can be though of as a LOOCV containment of all points (what is the minimum
radius to do so).
}
