% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{inner_containment_conformal_score_mode_radius_1d}
\alias{inner_containment_conformal_score_mode_radius_1d}
\title{Helper function for simulation-based conformal score calculation based
potentially based on mode clustering and changing radius values.}
\usage{
inner_containment_conformal_score_mode_radius_1d(
  df_row_group,
  simulations_group_df,
  data_column_names,
  simulation_info_df,
  list_radius_info,
  list_grouping_id,
  verbose = FALSE
)
}
\arguments{
\item{df_row_group}{data frame with new point information to be assessed}

\item{simulations_group_df}{data frame with multiplee simulated points}

\item{data_column_names}{character vector of column names that define
locational information of points}

\item{simulation_info_df}{a dataframe with information of each simulation's
psuedo-density estimate, mode clustering, ranking of psuedo-density (within
cluster and overall). We expect a \code{psuedo_density}, \code{groupings}
\code{ranking} and  \code{group_ranking} columns. See
\code{EpiCompare::inner_expanding_info} for expected structure.}

\item{list_radius_info}{list of lists of radius information per each mode.}

\item{list_grouping_id}{list of vectors of indices (grouped by mode cluster)
and ordered by psuedo-density values. These values are relative to the row
indices in simulation_info_df}

\item{verbose}{boolean, if progress should be tracked with a progress bar}
}
\value{
data.frame with a row for new point with a
section column \code{containment_val} which is the discrete simulation-based
conformal score. If \code{df_row_group} only contained location columns in
\code{data_column_names} then a new column named \code{index} is also
included, and is related to the row numbers/names of \code{df_row_group}.
}
\description{
Helper function for simulation-based conformal score calculation based
potentially based on mode clustering and changing radius values.
}
