% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{inner_discrete_mass_cde}
\alias{inner_discrete_mass_cde}
\title{inner function for estimation of probability of mass with CDE above or equal
to value}
\usage{
inner_discrete_mass_cde(cde_vec, delta_y = 16/99)
}
\arguments{
\item{cde_vec}{vector of CDE values for Y=y|X=x}

\item{delta_y}{space between each y value}
}
\value{
mass vector
}
\description{
inner function for estimation of probability of mass with CDE above or equal
to value
}
