% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{inner_f}
\alias{inner_f}
\title{f function for data generation for example in Lei and Wasserman (2014) paper}
\usage{
inner_f(x)
}
\arguments{
\item{x}{x value (can be a vector)}
}
\value{
f(x)
}
\description{
\eqn{f(x) = (x-1)^2(x+1)}
}
