% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{inner_g}
\alias{inner_g}
\title{g function for data generation for example in Lei and Wasserman (2014) paper}
\usage{
inner_g(x)
}
\arguments{
\item{x}{x value (can be a vector)}
}
\value{
g(x)
}
\description{
\eqn{g(x) = 2 sqrt(x+.5) II(x >= -.5)}{
     g(x) = 2\sqrt(x+.5)\mathbb{I}(x\geq -.5)}
}
