% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{inner_mixture}
\alias{inner_mixture}
\title{inner mixture creation function for \eqn{Y|X = x} for data generation for
example in Lei and Wasseerman (2014) paper}
\usage{
inner_mixture(x, group)
}
\arguments{
\item{x}{x value (can be a vector)}

\item{group}{group value (1,2) which mixture the y is from  (can be a vector)}
}
\value{
mixture value f(x) - g(x) or f(x) + g(x)
}
\description{
\deqn{f(x) + [-1,1](\code{group} == 1) g(x)}
}
\details{
where
\enumerate{
\item \eqn{f(x) = (x-1)^2(x+1)},
\item \eqn{
    g(x) = 2 sqrt(x+.5) II(x >= -.5)}{g(x) = 2\sqrt(x+.5)\mathbb{I}(x\geq -.5)},
}
}
