% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{inner_sigma}
\alias{inner_sigma}
\title{sigma function for data generation for example in Lei and Wasserman (2014)
paper}
\usage{
inner_sigma(x)
}
\arguments{
\item{x}{x value (can be a vector)}
}
\value{
sigma(x)
}
\description{
\deqn{sigma(x) = 1/4 + |x|.}{\sigma(x) = 1/4 + |x|.}
}
