% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{inner_sigma_zero}
\alias{inner_sigma_zero}
\title{0- sigma function for data generation for example in Lei and Wasserman (2014)
paper. To help just return the mean lines}
\usage{
inner_sigma_zero(x)
}
\arguments{
\item{x}{x value (can be a vector)}
}
\value{
sigma_zero(x)
}
\description{
\deqn{sigma(x) = .0001}{\sigma(x) = .0001}
}
