% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lei_wasserman.R
\name{lei_wasserman_data}
\alias{lei_wasserman_data}
\title{Generate data from example in Lei and Wasseerman (2014) paper}
\usage{
lei_wasserman_data(n = 1000, sigma_function = inner_sigma)
}
\arguments{
\item{n}{number of observations}

\item{sigma_function}{a function for the sigma function. The default is
described above, but using \code{function(x){0.0001}} would return the means
of the functions}
}
\value{
data.frame for (x,y) pairs of \code{n} simulations from the
generative process
}
\description{
The generative model is defined as
\deqn{X ~ Unif(-1.5, 1.5)}{X \sim Unif(-1.5, 1.5)}
\deqn{(Y|X=x) ~ .5 N(f(x)-g(x), sigma^2(x)) + .5 N(f(x)+g(x),
      sigma^2(x))}{(Y|X=x) \sim .5 N(f(x)-g(x),
       \sigma^2(x)) + .5 N(f(x)+g(x), \sigma^2(x))}
where
\enumerate{
\item \eqn{f(x) = (x-1)^2(x+1)},
\item \eqn{g(x) = 2 sqrt(x+.5) II(x >= -.5)}{
g(x) = 2\sqrt(x+.5)\mathbb{I}(x\geq -.5)}, and
\item \eqn{sigma(x) = 1/4 + |x|.}{\sigma(x) = 1/4 + |x|.}
}
}
