% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{maxmin_distance_vector}
\alias{maxmin_distance_vector}
\title{Calculate minimum radius for each group to contain the true data points.}
\usage{
maxmin_distance_vector(
  truth_df,
  simulations_grouped_df,
  data_column_names = c("S", "I", "R"),
  .all_info = F
)
}
\arguments{
\item{truth_df}{data.frame of true points (locations defined by
\code{data_column_names})}

\item{simulations_grouped_df}{grouped_df, with points grouped together}

\item{data_column_names}{column names for both the \code{truth_df} and
\code{simulations_grouped_df} that contain locations.}

\item{.all_info}{boolean, indicators what should be returned.}
}
\value{
If \code{.all_info} is \code{FALSE} then a data.frame of grouped
information and minimum distance per group in the
\code{simulations_grouped_df}. Else both that and a data.frame that details
the minimum covering distance for each point in the \code{truth_df} (by the
grouped set of points) is given. Note the order of the grouping index when
using these data.frames.
}
\description{
Fast version.
}
