% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{maxmin_distance_vector_old}
\alias{maxmin_distance_vector_old}
\title{Calculate minimum radius for each group to contain the true data points.}
\usage{
maxmin_distance_vector_old(
  truth_df,
  simulations_grouped_df,
  data_column_names = c("S", "I", "R")
)
}
\arguments{
\item{truth_df}{data.frame of true points (locations defined by
\code{data_column_names})}

\item{simulations_grouped_df}{grouped_df, with points grouped together}

\item{data_column_names}{column names for both the \code{truth_df} and
\code{simulations_grouped_df} that contain locations.}
}
\value{
data.frame of grouped information and minimum distance per group
in the \code{simulations_grouped_df}.
}
\description{
Slow version.
}
