% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{maxmin_inner}
\alias{maxmin_inner}
\title{minimum distance to contain all points of first set with union of balls of
second set}
\usage{
maxmin_inner(df_row, df_col, only_val = T)
}
\arguments{
\item{df_row}{data.frame, rows are observations, all columns are used in
distance calculation.}

\item{df_col}{data.frame, rows are observations, all columns are used in
distance calculation.}

\item{only_val}{boolean, effects return of scalar (TRUE) or vector (FALSE)
as described in details.}
}
\value{
single minimum distance scalar or minimum distance vector
}
\description{
Inner function, faster version. For each row in \code{df_row}, the minimum
distance to a point in \code{df_col} is calculated and then the maximum of
this vector is returned. If \code{only_val} is \code{FALSE}, then the
minimum distance is returned, else the vector per each row of \code{df_row}'s
minimum distance to a point in \code{df_col} is returned.
}
