% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{mode_clustering_1d}
\alias{mode_clustering_1d}
\title{Find mode clusters of euclidean objects based on distance - densities}
\usage{
mode_clustering_1d(
  df_info,
  position,
  naming_info = NULL,
  sigma,
  maxT = 50,
  eps = 1e-05,
  diff_eps = 1e-05,
  verbose = TRUE
)
}
\arguments{
\item{df_info}{data frame with rows of observations to mode cluster}

\item{position}{character vector of columns that define the location of the
points}

\item{naming_info}{a data frame with information about the data points (same
number of rows as \code{df_info})}

\item{sigma}{density scalar / sigma value}

\item{maxT}{int, maximum number of iterations}

\item{eps}{float, difference between 2 sequential points for which to stop
walking toward the mode for a give point's walk}

\item{diff_eps}{float, if the final step of each of the points is within
this distance from each-other they will be grouped together.}

\item{verbose}{boolean, if this progression should be verbose}
}
\value{
data frame with \code{naming_info} and a \code{grouping} column which
indicates which mode group each observation is in.
}
\description{
Find mode clusters of euclidean objects based on distance - densities
}
