% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Y.R
\name{multimode_density_f}
\alias{multimode_density_f}
\title{obtain conditional pdf value for y values for "v" distribution}
\usage{
multimode_density_f(
  y,
  x_val,
  beta = c(-1, 1),
  intercept = c(0, 0),
  sd = c(1, 1),
  proportions = c(0.5, 0.5)
)
}
\arguments{
\item{y}{can be a vector}

\item{x_val}{a single value}

\item{right_beta}{vector length 2, see above}

\item{right_intercept}{vector length 2, see above}

\item{right_sd}{vector length 2, see above}

\item{right_proportions}{vector length 2, see above}
}
\value{
conditional pdf values for y vector
}
\description{
conditional pdf is defined under y ~ p1 M1(x) + p2 M2(x)
where (M1,M2)(x) ~ right_intercept + N(right_beta*x, right_sd) and
(p1,p2) = right_proportions
}
