% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Y.R
\name{multimodel_density_split_mixture}
\alias{multimodel_density_split_mixture}
\title{obtain conditional pdf value for y values for "Y" distribution}
\usage{
multimodel_density_split_mixture(
  y,
  x_val,
  right_beta = c(-1, 1),
  right_intercept = c(0, 0),
  right_sd = c(1, 1),
  right_proportions = c(0.5, 0.5)
)
}
\arguments{
\item{y}{can be a vector}

\item{x_val}{a single value}

\item{right_beta}{vector length 2, see above}

\item{right_intercept}{vector length 2, see above}

\item{right_sd}{vector length 2, see above}

\item{right_proportions}{vector length 2, see above}
}
\value{
conditional pdf values for y vector
}
\description{
conditional pdf is defined as following:
}
\details{
If x < -1 we have y ~ norm(mu = 0, sd = 1), else
y ~ p1 M1(x) + p2 M2(x)
where (M1,M2)(x) ~ right_intercept + N(right_beta*x, right_sd) and
(p1,p2) = right_proportions
}
