% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{my_pdist}
\alias{my_pdist}
\title{calculate distance between two different df's rows}
\usage{
my_pdist(df1, df2)
}
\arguments{
\item{df1}{data frame  (n, p)}

\item{df2}{data frame  (m, p)}
}
\value{
matrix of distances (n, m)
}
\description{
calculate distance between two different df's rows
}
