% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_funtional.R
\name{pos_int}
\alias{pos_int}
\title{Named after assuming positive integers, but not checked}
\usage{
pos_int(x)
}
\arguments{
\item{x}{vector of numerical values (or single value)}
}
\value{
subset of \code{x} that greater than 0.
}
\description{
Named after assuming positive integers, but not checked
}
