% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{psuedo_density_mode_cluster_1d}
\alias{psuedo_density_mode_cluster_1d}
\title{Mode clustering for euclidean data (distance based)}
\usage{
psuedo_density_mode_cluster_1d(
  X_mat,
  G_mat = X_mat,
  sigma,
  maxT = 30,
  eps = 1e-05,
  verbose = TRUE,
  list_out = FALSE
)
}
\arguments{
\item{X_mat}{matrix of data points (in rows), that define the kernel density}

\item{G_mat}{matrix of data points (in rows) that need to be "walked" to
their modes}

\item{sigma}{density scalar / sigma value}

\item{maxT}{int, maximum number of iterations}

\item{eps}{float, difference between 2 sequential points for which to stop
walking toward the mode for a give point's walk}

\item{verbose}{boolean, if this progression should be verbose}

\item{list_out}{if we should return every step of the process (for fun
visualization purposes and visual checks only)}
}
\value{
list of :
\itemize{
\item matrix of \code{G_mat} points walked up all the way
\item if \code{list_out} is \code{TRUE} then list of matrices for each step,
else \code{NULL}
}
}
\description{
Walks up a guassian kernel density estimate to find modes
}
\details{
Similar to more complex code developed by Yen-Chi Chen and Chris Genovese
}
