% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_prediction.R
\name{simulation_based_conformal_1d_complex}
\alias{simulation_based_conformal_1d_complex}
\title{global wrapper for simulation-based conformal score calculation that allows
for mode clustering and changes of radius for 1d examples}
\usage{
simulation_based_conformal_1d_complex(
  truth_df,
  simulations_df,
  smooth_functions = list(),
  data_column_names = c("y"),
  .maxT = 50,
  .sigma_string = "25\%",
  .diff_eps = 1e-06,
  .eps = 1e-06,
  return_min = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{truth_df}{data frame with new point information to be assessed}

\item{simulations_df}{data frame with multiplee simulated points}

\item{smooth_functions}{a list of functions a function for the smoothing
approach for the varied approach}

\item{data_column_names}{columns of \code{truth_df} and
\code{simulations_df} that correspond to the output space.}

\item{.maxT}{int, max number of steps for mode clustering mean-shift
algorithm}

\item{.sigma_string}{string, the quantile of the distance matrix to define
the sigma (e.g. \code{"30\%"})}

\item{.diff_eps}{float, the error allows between mode clustered final steps
to still be counted as the same mode.}

\item{.eps}{float, when to stop the model clustering (if maxT doesn't stop it
first)}

\item{return_min}{boolean, if list of information returned is mimimum (for slurm)}

\item{verbose}{boolean, be verbose about progression}
}
\value{
list of information...
}
\description{
global wrapper for simulation-based conformal score calculation that allows
for mode clustering and changes of radius for 1d examples
}
