% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformal_pdf_approach.R
\name{stepwise_conformal_cde_update}
\alias{stepwise_conformal_cde_update}
\title{stepwise transform of CDE to become comformal}
\usage{
stepwise_conformal_cde_update(
  n,
  prob_cde,
  lambda = -1,
  alpha = 0,
  monotonically_increasing = F,
  delta = 1e-12,
  scaling_constraint = F
)
}
\arguments{
\item{n}{number of conformal values (length of \code{prob_cde} - 1)}

\item{prob_cde}{vector of mass of probability (defined by fitted density)
between each conformal based level grouping}

\item{lambda}{optimization constraint (weight for smoothing)}

\item{alpha}{elasticnet penalty (recommend default at \code{0})}

\item{monotonically_increasing}{constraint to require scaling to increase
(recommend default at \code{FALSE})}

\item{delta}{stabilizing constant (added to diagonal if needed)}

\item{scaling_constraint}{boolean if we add a constraint to make sure that
the final density has mass 1 (currently set at \code{FALSE} - but not for
a good reason).}
}
\value{

}
\description{
Solving for the x in the following general equation
\deqn{\min_{x, z} 1/2(Sx-b)^T(Sx-b) +
      \lambda (\alpha z^Tz + (1-\alpha) 1^Tz)}
subject to some of the following equations
\tabular{rl}{
\eqn{A_1}: & \eqn{z_i \geq x_i - x_{i+1}}  \\
\eqn{A_2}: & \eqn{z_i \geq - x_i + x_{i+1}}  \\
\eqn{A_3}: & \eqn{x_i \geq 0}  \\
\eqn{A_4}: & \eqn{x_i - x_{i+1} \geq 0}  \\
}
}
\details{
With \eqn{A_1, A_2, A_3} constraints for \code{monotonically_increasing} is
\code{FALSE} and \eqn{A_1, A_3, A_4} if \code{monotonically_increasing} is
\code{TRUE}.
}
