% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_diagnostics.R
\name{vis_stepwise_scaling_only}
\alias{vis_stepwise_scaling_only}
\title{visualize scaling values}
\usage{
vis_stepwise_scaling_only(stepwise_scaling, x = 1:length(stepwise_scaling))
}
\arguments{
\item{stepwise_scaling}{scaling values (for each level)}

\item{x}{x axis values, e.g. the level set values, proportion of mass
values, etc. The default is just an index.}
}
\value{
\code{ggplot} visual
}
\description{
visualize scaling values
}
